+++ .gitignore
@@ -21,6 +21,7 @@ tables.xlsx
 cfru_thread.txt
 src/Tables/criminon_trainer_spreads.c
 src/start_menu_unbound.c
+src/tier_bans.c
 src/unbound_util.c
 strings/unbound.string
 Cover Art.png
