+++ assembly/data/species_tables.json
@@ -95,6 +95,7 @@
 		"SPECIES_ETERNATUS",
 		"SPECIES_ETERNATUS_ETERNAMAX",
 		"SPECIES_ZARUDE",
+		"SPECIES_ZARUDE_DADA",
 		"SPECIES_KUBFU",
 		"SPECIES_URSHIFU_SINGLE",
 		"SPECIES_URSHIFU_RAPID",
@@ -236,6 +237,8 @@
 		"SPECIES_ARCEUS_DRAGON",
 		"SPECIES_ARCEUS_DARK",
 		"SPECIES_ARCEUS_FAIRY",
+		"SPECIES_CALYREX_ICE_RIDER",
+		"SPECIES_CALYREX_SHADOW_RIDER",
 		"SPECIES_CINDERACE",
 		"SPECIES_DARKRAI",
 		"SPECIES_DARMANITAN_G",
@@ -264,6 +267,8 @@
 		"SPECIES_LUGIA",
 		"SPECIES_LUNALA",
 		"SPECIES_MARSHADOW",
+		"SPECIES_MAGEARNA",
+		"SPECIES_MAGEARNA_P",
 		"SPECIES_MEWTWO",
 		"SPECIES_NAGANADEL",
 		"SPECIES_NECROZMA_DUSK_MANE",
@@ -863,8 +868,7 @@
 		"SPECIES_DRAGONITE",
 		"SPECIES_MEW",
 		"SPECIES_SHEDINJA",
-		"SPECIES_KARTANA",
-		"SPECIES_SLOWKING_G"
+		"SPECIES_KARTANA"
 	],
 
 	"smogonAverageMonsBan": [
@@ -878,6 +882,7 @@
 	"smogonScalemonsBan": [
 		"SPECIES_ABRA",
 		"SPECIES_CARVANHA",
+		"SPECIES_DARMANITAN_G",
 		"SPECIES_GASTLY",
 		"SPECIES_SHEDINJA"
 	],
@@ -1260,6 +1265,7 @@
 		"SPECIES_ETERNATUS",
 		"SPECIES_ETERNATUS_ETERNAMAX",
 		"SPECIES_ZARUDE",
+		"SPECIES_ZARUDE_DADA",
 		"SPECIES_KUBFU",
 		"SPECIES_URSHIFU_SINGLE",
 		"SPECIES_URSHIFU_RAPID",
@@ -1403,6 +1409,7 @@
 		"SPECIES_ETERNATUS",
 		"SPECIES_ETERNATUS_ETERNAMAX",
 		"SPECIES_ZARUDE",
+		"SPECIES_ZARUDE_DADA",
 		"SPECIES_KUBFU",
 		"SPECIES_URSHIFU_SINGLE",
 		"SPECIES_URSHIFU_RAPID",
