+++ hooks
@@ -755,4 +755,5 @@ CanSpeciesEvolve 801D874 1
 IsSpeciesBannedFromRandomizer 801D87C 1
 GetBaseStatsTotal 801D884 1
 IsAutoScrollEnabled 801D88C 0
+TierBanScreenFromScript 801D894 0
 #endif
