+++ include/new/frontier.h
@@ -44,6 +44,9 @@ bool8 AreMegasZMovesBannedInTier(u8 tier);
 bool8 IsMegaZMoveBannedBattle(void);
 bool8 IsMoveBannedInRingChallenge(u16 move, u8 bank);
 bool8 IsMoveBannedInRingChallengeByMon(u16 move, struct Pokemon* mon);
+bool8 PokemonTierBan(const u16 species, const u16 item, const struct BattleTowerSpread* const spread, const struct Pokemon* const mon, const u8 tier, const u8 checkFromLocationType);
+bool8 IsMonBannedInTier(struct Pokemon* mon, u8 tier);
+bool8 IsSpeciesBannedInTier(u16 species, u16 tier, u16 battleFormat);
 u16 GetCurrentBattleTowerStreak(void);
 u16 GetBattleMineStreak(u8 type, u8 tier);
 u16 GetRingChallengeSteak(u8 type);
