+++ strings/frontier_records.string
@@ -73,6 +73,9 @@ No Restrictions
 #org @gText_SmogonGen7OU
 Smogon OU
 
+#org @gText_SmogonOUDoubles
+Smogon Doubles OU
+
 #org @gText_SmogonGen7Uber
 Smogon Ubers
